requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
pytest==7.1.2
requests==2.27.1
selenium==4.1.5
pytest-xdist==2.5.0

